requirements.txt
streamlit>=1.32.0
google-genai>=0.2.0
numpy>=1.24.0
